refactor(helpers): remove helper index file due to conflicts.
There are also some import efficiency concerns with this approach.
Loading a single helper was causing all helpers to be loaded.
