--- helpers/index.js
@@ -1,35 +1,0 @@
-import convertObjectToQueryParams from './convertObjectToQueryParams'
-import activateZenDesk from './activateZenDesk'
-import connect from './connect'
-import convertSlugToUUID from './convertSlugToUUID'
-import convertStringToSlug from './convertStringToSlug'
-import createAlertObject from './createAlertObject'
-import deepMergeJSONAPIObjectCollections from './deepMergeJSONAPIObjectCollections'
-import findAndEvaluateModifiers from './findAndEvaluateModifiers'
-import formatGameString from './formatGameString'
-import getBase64FromFileInput from './getBase64FromFileInput'
-import isUUID from './isUUID'
-import isRequired from './isRequired'
-import parseJSONAPIResponseForEntityType from './parseJSONAPIResponseForEntityType'
-import pushTrackableEventToDataLayer from './pushTrackableEventToDataLayer'
-
-
-
-
-
-export {
-  activateZenDesk,
-  connect,
-  convertObjectToQueryParams,
-  convertSlugToUUID,
-  convertStringToSlug,
-  createAlertObject,
-  deepMergeJSONAPIObjectCollections,
-  findAndEvaluateModifiers,
-  formatGameString,
-  getBase64FromFileInput,
-  isUUID,
-  isRequired,
-  parseJSONAPIResponseForEntityType,
-  pushTrackableEventToDataLayer,
-}
