+++ store/reducers/users.js
@@ -8,10 +8,8 @@ import Cookies from 'js-cookie'
 // Component imports
 import actionTypes from '../actionTypes'
 import initialState from '../initialState'
-import {
-  deepMergeJSONAPIObjectCollections,
-  parseJSONAPIResponseForEntityType,
-} from '../../helpers'
+import deepMergeJSONAPIObjectCollections from '../../helpers/deepMergeJSONAPIObjectCollections'
+import parseJSONAPIResponseForEntityType from '../../helpers/parseJSONAPIResponseForEntityType'
